File organization update
-moved all source code files into the new src directory
-updated makefile with correct locations of all source code files
-added CODEOWNERS file to designate code owernership
