+++ CODEOWNERS
@@ -1,0 +1,2 @@
+# Define code owners for all files in the repo
+*     @MikeHeiber
