+++ src/Lattice.cpp
