+++ src/Lattice.h
